Create List<File> interface class for keychain lookup list operations
To allow a more java native way of handling the keychain lookup list I
wrapped the `SecurityUtil` methods in a custom class
`KeychainLookupList`. The `KeychainLookupList` class implements the
`List<File>` interface and allows a way better access with a common API.

This commit also adds a test spec for the new class which also tests
through the `KeychainLookupList` the static methods in `SecurityUtil`.
In future commits I might move more and more logic from the static
methods into the `KeychainLookupList` class to remove the utils class
again.