Pass loglevel from main gradle project to the sub projects
Description
===========

If you choose to build the whole unity project with `--info` loglevel
than this log setting isn't passed down to the exported sub-projects.
This patch adds this functionality by reading the log level from the
`StartParameter` object and setting the matching commandline switches.
The only supported levels are:

* debug
* info
* quiet

Changes
=======

![FIX] log-level settings passed to exported sub projects