Fix up-to-date check for task :buildKeychain
Description
===========

The `up-to-date` check for the `:buildKeychain` task always returned
`false` because gradle believed the output file wasn't generated by
gradle. I made a simple mistake in the `OutputFiles` definition for the
task. I used a `FileTree` to create the `FileCollection` which yields a
size `0` list when building a clean build. Next build would delete the
output file because gradle never recorded it as a output file. So this
pattern happens over and over again with each rebuild.

I switched from `project.fileTree` to `project.files` and pass in a
`File` object. I can't remember why I used the fileTree in the first
place. I also changed that for `:importProvisioningProfiles`. I added a
testspec for the keychain task but not yet for the provisioning profiles
one because of `fastlane`.

Changes
=======

![FIX]![IOS] `up-to-date` check for task `:buildKeychain`