Fix up-to-date check for export task when project version changes
Description
===========

During task execution of `UnityBuildPlayerTask` aka
`export${Platform}${Environment}, the version passed to unity was
fetched from the project. If the project version changes the task has no
tracking of the old value so it things the task is still `up-to-date`.

This patch adds a new `Input` property called `version` to the
`UnityBuildPlayerTask` class and maps the default value to
`project.version`.

Changes
=======

![FIX] `UnityBuildPlayerTask` **up-to-date** check.