Fix version to string conversion crash
Description
===========

The `project.version` property is by default an `Object` in most cases
it will be a `String` value. I changed the setting this `version` value
from a direct assignment into a convention mapping. What is important to
note is that instead of the original `getter` the mapped `closure` will
be executed instead. I added a lot of conversion methods to convert an
object into a `String`. This wasn't done in the mapping. I added a very
quick regression test in `UnityBuildPluginIntegrationSpec`. I wanted to
do the same for the iOS plugin but this is impossible at the moment due
to the untestable state of the plugin.

Changes
=======

![FIX] `project.version` toString conversion