## Description
This patch introduces the gradle [Provider API] to
`net.wooga.build-unity` to enable future proof lazy configuration.

### Breaking Changes

The change to the [Provider API] comes with some breaking changes. I
made the deliberate choice not to add backwards compatible API methods
to see what the base API can offer out of the box. It also makes the
change to this new and cleaner API more clear.

### Noticeable example

The [Provider API] is in it's core very `functional`. The provider
objects are lazy and need to be queried for a value. As in every new API
we need to bind legacy API to the new system. This is what happens here:

**DefaultUnityBuildPluginExtension.groovy**
```groovy
appConfigsDirectory.set(assetsDir.dir(UnityBuildPluginConsts.DEFAULT_APP_CONFIGS_DIRECTORY))

assetsDir.set(project.provider(new Callable<Directory>() {
    @Override
    Directory call() throws Exception {
        UnityPluginExtension unity = project.extensions.getByType(UnityPluginExtension)
        def assetDir = project.layout.directoryProperty()
        assetDir.set(project.file(unity.assetsDir))
        assetDir.get()
    }
}))
```

The `assetsDir` property provided by `net.wooga.unity` plugin is not
implemented as a `Property<Directory>`. So we need to wrap it in a
custom `Provider` to make this value usable in a `Provider` chain.

The API has still some overhead and boilerplate code but makes it way
easier to construct lazy evaluated properties.

### Property API changes

As of this moment, gradle doesn't support auto generation of property
methods. I removed these API calls along with the old setter. It is
possible to bring these back.

```groovy
build-unity {
  toolsVersion "1.1.1"      //no longer working
  toolsVersion = "1.1.1"    //only property access enabled
  toolsVersion.set("1.1.1") //native way to set a `Provider` property
}
```
I had to change a lot of tests because I normaly use the short property
set methods.

### Property ambiguous type setters

The [Provider API] also breaks with the old API habit to set values as
`Object`s and evaluate in the getter. This old style API was used to
support property setter with multiple types. A `File` property could be
set with a `File`, `String` or `Callable` object. This is not supported
out of the box and needs custom set methods.

```groovy
build-unity {
  appConfigsDirectory = file('path/to/something')              //only accepted set method
  appConfigsDirectory = 'path/to/something'                    //no longer provided
  appConfigsDirectory = {'path/to/something'}                  //no longer provided because not needed

  appConfigsDirectory.set(file('path/to/something'))           //native way to set
  appConfigsDirectory.set(provider{file('path/to/something')}) //native way to set a lazy value
}
```

## Questions

* do we want the quick set methods
* do we want the ambiguous type setters

## Changes

* ![CHANGE] supported gradle version from `4.0` - `4.10`
* ![BREAK]  `UnityBuildPluginExtension` API
* ![BREAK]  `GradleBuild` task API
* ![BREAK]  `UnityBuildPlayerTask` task API
* ![ADD]    Provider API to `UnityBuildPluginExtension`
* ![ADD]    Provider API to `GradleBuild`
* ![ADD]    Provider API to `UnityBuildPlayerTask`

[Provider API]: https://docs.gradle.org/current/userguide/lazy_configuration.html
[NEW]:https://atlas-resources.wooga.com/icons/icon_new.svg "New"
[ADD]:http://resources.atlas.wooga.com/icons/icon_add.svg "Add"
[IMPROVE]:http://resources.atlas.wooga.com/icons/icon_improve.svg "IMPROVE"
[CHANGE]:http://resources.atlas.wooga.com/icons/icon_change.svg "Change"
[FIX]:http://resources.atlas.wooga.com/icons/icon_fix.svg "Fix"
[UPDATE]:http://resources.atlas.wooga.com/icons/icon_update.svg "Update"

[BREAK]:http://resources.atlas.wooga.com/icons/icon_break.svg "Remove"
[REMOVE]:http://resources.atlas.wooga.com/icons/icon_remove.svg "Remove"
[IOS]:http://resources.atlas.wooga.com/icons/icon_iOS.svg "iOS"
[ANDROID]:http://resources.atlas.wooga.com/icons/icon_android.svg "Android"
[WEBGL]:http://resources.atlas.wooga.com/icons/icon_webGL.svg "Web:GL"