Fix integration test build setup (#24)
Description
===========

This patch fixes some minor issues with the integration test setup. We
need to disable the gradle 5 compatibility warnings as all tests fail
because of it. There is also a setup bug for the provided build targets
in the mock app configs. We should only test with valid build target
values.

Changes
=======

![FIX] test values for build target in integration test
![IMPROVE] disable gradle 5 compatibility warnings