Fix double buildTarget argument when exporting with default target set (#25)
Description
===========

When using the `UnityBuildPlayerTask` within a gradle project that also
configured the `defaultBuildTarget` value, then the
`UnityBuildPlayerTask` might configure the `-buildTarget` flag for the
batchmode invocation two times. It's undefined for me if the first or
second value is used by Unity. This patch adds a test spec and the
necessary patch to fix this issue.

Changes
=======

![FIX] double buildTarget argument when exporting with default target configured