Improve Linux support
Description
===========

The latest patches to this package did not include any checks or
implementations for linux. As unity moves the linux editor to production
it makes sense to gain better support for this OS as well.
As a test setup I added the same dockerfile we also use in the other
unity related plugins and libraries.

Changes
=======

![IMPROVE] linux support