Improve inner task execution by providing gradle version to execute with
Description
===========

For newer Unity version it is desired that the exported gradle project
runs with a configured gradle version. At the moment, gradle uses the
same version as the root project version to execute the build in the
exported projects. This patch allows to set a custom gradle version in
the `appConfig`. Gradle will fall back to the root gradle version if the
property `gradleVersion` is not defined.

Changes
=======

![IMPROVE] configuration by providing gradle version for exported task runner