Delete unused webGL plugin
Description
===========

This patch removes the first attempt of a webGL build plugin. There are
to many open questions still and the `brotli` dependency crawls over to
the other plugins as well. I don't want to change the project setup just
now or split the sources to limit down these effects. Since we don't use
this plugin on production code it maybe is cleaner to just delete it
again.

Changes
=======

![BREAK] delete unused `webGL` plugin