Add provisioningName property to ImportProvisioningProfile (#40)
Description
===========

Fastlane `sigh` tries to find a matching provisioning profile based on
the `team_id` and `app_id` parameters. There could be multiple valid
profiles matching these criteria and fastlane will by default download
the first one it find. There is an option to download all matching
profiles but I opted to not support this switch for now because this
would mean the task has to generate multiple output files as well which
would be a bigger change.
There is another filter switch `--provisioning_name`. With this switch
set, `sigh` tries to find a matching profile with exactly that name.

Test-ability changes
--------------------

The last patch added basic test setup for the
`ImportProvisioningProfile` task class. This patch doubles down on it
and provides a matrix test for all task input properties. It only checks
the invocation of `fastlane`. So this is not a real integration test.

Changes
=======

![ADD] `provisioningName` property to `ImportProvisioningProfile`