Add iOS adhoc support
Description
===========

To import _adhoc_ provisioning profiles with `fastlane` correctly in
case of _adhoc_ builds we need to tell fastlane that we want an _adhoc_
profile. Fastlane will fail and try to generate a _store_ or
_enterprise_ profile instead.

This patch adds a new property `adhoc` to `ImportProvisioningProfile`
and `IOSBuildPluginExtension` which will pass this information over to
fastlane. It would had been possible to set `SIGH_AD_HOC` to make this
clearer in both useage and output.

Be aware!
=========

Because of the nature of boolean values (true/false there is no notion
of not set) this new logic will always set the `--adhoc` field. When we
swich to the gradle provider API this issue will be fixed since a gradle
property has the information of `present` or not. This also means that
the environment variable `SIGH_AD_HOC` will have no effect anymore.

Changes
=======

* ![ADD] ![IOS] `adhoc` support