Add 'commitHash' property to UnityBuildPlayerTask (#44)
Description
===========

A long with the `toolsVersion` and gradle project version we want to
know the current `commitHash` in Unity without quering git.
This patch adds a new property `commitHash` to the
`UnityBuildPlayerTask` class. The property is configured similar as the
`toolsVersion` property. There is a fallback to the
`UNITY_BUILD_COMMIT_HASH` environment variable and
`unityBuild.commitHash` gradle property

Changes
=======

* ![ADD] 'commitHash' property to `UnityBuildPlayerTask`