Improve iOS keychain task error messages (#45)
Description
===========

The `buildKeychain` task for iOS projects uses the `security`
commandline tool to setup a new keychain and import optional
certificates. This operation is handled in the _interactive_ mode to
spawn only one `security` process. In case of an error we only show a
message that `security` failed but not exactly what failed.
I changed the task to register both `stderr` and `stdout` and print both
to the gradle logger. In additon to this I also handle the exit value
and throw a `ExecException` with the `stderr` string as message. This
should result in a clearer error message what happend.
As an example. If the user provides a wrong certificate (.p12)
passphrase the gradle process would end like:

```
FAILURE: Build failed with an exception.
* What went wrong:
Execution failed for task ':buildKeychain'.
> security: SecKeychainItemImport: MAC verification failed during PKCS12 import (wrong password?)
  import: returned 1
```

Changes
=======

* ![IMPROVE] ![IOS] keychain task error messages