Refactor external build cache setup and configuration (#50)
* Refactor external build cache setup and configuration

Description
===========

Exporting and building an exported gradle project from unity will at the
moment move the `build` directory of all projects of the extported
project to a custom location. As an example:

_a project exported to `build/export/ios/project`_ will have its gradle
cache dir and build dir set to `build/export/ios/buildCache`.

This was achieved by loading a gradle init script from the current
project directory named `exportInit.gradle`. This script needed the
following content and was *not* part of this plugin:

```groovy
println "------------------------------------------------------------"
println "                        INIT                                "
println "------------------------------------------------------------"

projectsLoaded {
    def buildDirBase = rootProject.file(rootProject.properties.get("export.buildDirBase", rootProject.projectDir))

    allprojects(new Action<Project>() {
        @Override
        void execute(Project project) {
            def newBuildPath = new File(buildDirBase, "${project.path.replace(':','/')}/build")
            println("set buildDir for project ${project.name} to $newBuildPath")
            project.buildDir = newBuildPath
        }
    })
}
```

The main purpose of this script is to set the `buildDir` folder for all
projects of a gradle multiproject build. This all worked well but was
rather unintuitiv and the fact that the main part, the init script
itself, wasn't a part of the plugin, makes the setup difficult.

This patch changes the implementation and streamlines the process.
The `GradleBuild` task and the `UnityBuildPluginExtension` have some new
properties:

- `exportInitScript` / `initScript`
- `cleanBuildDirBeforeBuild`
- `exportBuildDirBase` / `buildDirBase`

There is no need for a custom `exportInit.gradle` in the project. The
plugin contains this script and will use it based on the values of the
new properties. It's possible now to provide an additional init script
and a way to let gradle clean the build dir before building.

_build.gradle_

```groovy

plugins {
   id 'net.wooga.build-unity' version 'x.x.x'
}

unityBuild {
   cleanBuildDirBeforeBuild = true
   exportBuildDirBase = new File("../buildCache")
}

```

Changes
=======

* ![IMPROVE] external build cache setup and configuration

* Fix usecase for relative paths to external base build dir

* Fix missing default setting for cleanBuildDirBeforeBuild

* Set custom init script mutal exclusive to build in script