Add testflight publish
Description
===========

This patch adds a new task type `PublishTestFlight` to the plugin. This
task uses `fastlane pilot upload`. This solution would bring the most
options for the future but I'm still unsure if we should not simple use
`altool`. The plugin creates a new task `publishTestFlight` and hooks
are the base information like fastlane credentials and path to the `ipa`
being published.

The publish task nows a few more parameters that are passed down to
`fastlane pilot upload`

| parameter | fastlane flag | description |
| --------- | ------------- | ----------- |
| appIdentifier | --app_identifier | STRING The bundle identifier of the app to upload or manage testers
| teamId | --team_id | The ID of your App Store Connect team if you're in multiple teams
| teamName | --team_name | STRING The name of your App Store Connect team if you're in multiple teams
| itcProvider | --itc_provider | STRING The provider short name to be used with the iTMSTransporter to identify your team. This value will override the automatically detected provider short name. |
| devPortalTeamId | --dev_portal_team_id | The short ID of your team in
the developer portal, if you're in multiple teams. Different from your
iTC team ID! |
| username | --username | STRING Your Apple ID Username
| skipSubmission | --skip_submission | [VALUE] Skip the distributing
action of pilot and only upload the ipa file |
| skipWaitingForBuildProcessing | --skip_waiting_for_build_processing | [VALUE] Don't wait for the build to process. If set to true, the changelog won't be set,
`distribute_external` option won't work and no build will be distributed to testers. (You might want to use this option if you are using this action on
CI and have to pay for 'minutes used' on your CI plan)
| ipa | --ipa | STRING     Path to the ipa file to upload
| password | | Your Apple ID Password

Other options might follow in the future but these should be enough for
a simple upload and testflight app setup.

The `publishTestFlight` task is wired to the global `publish` task when
the extention parameter `publishToTestFlight` is set to true. (default
false)

Changes
=======

![ADD] ![IOS] task `PublishTestflight`
![IMPROVE] test setup
![IMPROVE] test coverage