Add versionCode property (#57)
Description
===========

This patch adds main support for a second version specifier the
`versionCode`. The `versionCode` is a `String` property even though
android uses an integer.

1. To pass it down to unity it has to be converted to `String` anyways
2. iOS has different rules for the [`CFBundleVersion`] which is a `String`

The default value is `null` and won't be send to Unity. The property can
be set on the `UnityBuildPlayerTask` or in the `buildUnity` extension.
In the later case all tasks of type `UnityBuildPlayerTask` will have
the same value set. It is also possible to set the value via environment
or gradle properties. I also made the `version` property of the
extension overrideable via environment or gradle properties. The main
fallback is still the project version. Only if this value for
project.version is `unspecified` will the extension check the environment/gradle
properties.

| property              | gradle property name      | environment variable      |
| --------------------- | ------------------------- | ------------------------- |
| version               | `unityBuild.version`      | `UNITY_BUILD_VERSION`     |
| versionCode           | `unityBuild.versionCode`  | `UNITY_BUILD_VERSION_CODE |

Changes
=======

* ![ADD] `versionCode` property

[CFBundleVersion]: https://developer.apple.com/documentation/bundleresources/information_property_list/cfbundleversion