Add secrets resolver base setup (#58)
Description
===========

This is a pretty huge patch and implements the basic tools needed to
fetch and pass secrets to Unity. The patch [#105] is introducing a new
configuration field called `secrets` to the `AppConfig`. The fields
intention is to declare all secretIds needed byt the `AppConfig`.
This patch implements the basic workflow to read the secretsIds, fetch
the secrets through a `SecretResolver` and pass them to Unity via
environment variables (this is the plan for the first implementation).

FetchSecrets
------------

I added a new task type `FetchSecrets` and configured tasks for all
found `AppConfig` files in the `AppConfig` lookup directory. This means
every `UnityBuildPlayerTask` generated from the plugin will have a
corresponding `FetchSecrets` task as a direct dependency. (I might
change this setup so any task of `UnityBuildPlayerTask` gets a fetch
secret task assigned)

The `FetchSecrets` task will read the list of configured secrets in the
`AppConfig` and pass them to a configured resolver of type
`SecretResolver`. There is no default implementation available at the
moment. Only the integration tests provide some mock resolvers. If no
resolver is set or no secret ids are configured the task will create an
empty output.

The output of the `FetchSecrets` file is a serialized yaml version of
an object the new helper class type `Secrets`. This class handles
unencrypted secrets and encrypts them before serializing to disk.

e.g.

```yml
!secrets
secrets:
  net.wooga.testCredential: !encryptedSecretText
    secretValue: xllADGqYIK5h3go11IMfx8Iv0eH6UHhukv9sH6VrO4bggTXU+54HUhjR/29Bojvq
  net.wooga.testCredential2: !encryptedSecretText
    secretValue: AZ5nk3W6sogqFz/w82iZkA4pcGHKEmZpWbM7Cvgd+xsJCtEL6sTP+UGtUDaDh0dt
  net.wooga.testCredential3: !encryptedSecretFile
    secretValue: !!binary |-
      o+nF/o+R7jTy9q5aLfUsJ0WjXVl5hwVuykjr/G9zJVtweA7e+c/1HEjkR4f3988x
  net.wooga.testCredential4: !encryptedSecretText
    secretValue: 11TjUVMnoVHF3v2G2H1l1jDLo7AvpoNbvvqy6Ck3Nbt/Yvk3gXKFKfBeHIN/MPF7
```

Both `Secrets` and `SecretResolver` build on top of other helper types
to describe resolved secrets (`Secret`, `EncryptedSecret`, and more)

UnityBuildPlayerTask
--------------------

Before execution of the Unity batchmode task the build task will decrypt
the secrets from disk and format them in a way suitable to be passed as
environment variables.

SecretStrings are passed as is. Secret files are passend in memory as
byte arrays and are saved to a random temp file before execution. The
temp files will be delted after the Unity job is finished (successfull
or not).

SecretKey
---------

To encrypt and decrypt the logic uses a `AES` key. This key is by
default generated at runtime. This means that cached secret files are
unreadable after the gradle build is done. For local machines it is
possible to configure a static key outside of the project. This part of
the implementation is not yet final. The patch already includes a setter
for the key but no way to generate a key or save it securely on the host
computer.

What is needed next are a real `SecretResolver` implementation which can
fetch secret strings and files from a secure vault.

Changes
=======

* ![ADD] Secret resolver base setup

[#105]:		https://github.com/wooga/wdk-unity-UnifiedBuildSystem/pull/105