Allow appConfig lookup key to be configurable
Description
===========

To keep the connection between this plugin and the internal library as
loose as possible I changed the appConfig lookup key to be configurable.
The default value is the new key used by the library `secretIds`. The
key can be changed to any string value in the task or for all
fetchSecret task via the plugin extension

Changes
=======

* ![CHANGE] make appConfig secret ids lookup key configurable