Add pods and workspace support
Description
===========

This patch adds a new task `podInstall` to the plugin which is a
dependend task for `xcodeArchive`. It only executes when a `Podfile` is
in the project directory. The `xcodeArchive` task has a new workspace
property which is set by default to the workspace generated by
podInstall. The workspace has a higher precedance than the project. If
the workspace does not exist the task falls back to the project.

The issue is the very nature how cocoapods handles workspaces. You can
configure pods to use a preexisting workspace or it generates a new
workspace based on the project it finds in the same directory.

The current patch only tries to solve a current workflow issue and needs
changes and tests!

Changes
=======

* ![ADD] `pods` and `workspace` support