Change FetchSecrets dependency to AppConfig
Description
===========

This patch is a preparation to split the secrets handling code from this
plugin. The `FetchSecrets` task has a property `appConfigFile` which
contains a list of `secretIds` to fetch. To make the task more generic I
removed these properties and replaced it with a new property:
`final ListProperty<String> secretIds` and provided the implementation
details where these secretsIds are coming from in the `UnityBuildPlugin`
class. This allows me to split out this task into it's own plugin
without ties to UBS.

Changes
=======

* ![CHANGE] `FetchSecrets` task AppConfig dependency