Move secrets handling implemenation in external gradle plugin
Description
===========

I created a new upstream plugin [net.wooga.secrets] which is a copy of
the secrets resolver and other types created in this project. The
package names have changed during the move but the functionality is 1 to
1 the same. This is only the first patch of multiple. This repo uses
unstable API's from internal packages. This should be removed/updated
over time.

Changes
=======

* ![CHANGE] extract secrets handling into external plugin [net.wooga.secrets]

[net.wooga.secrets]:   https://github.com/wooga/atlas-secrets