Add support for custom arguments for the build player task
Description
===========

The `UnityBuildPlayerTask` provides special UBS arguments with a special
commandline flag `-CustomArgs` and a single long value with multiple
arguments enoded as "key=value;". There was no dynamic way of adding
values to this flag. This patch adds a new property `customArguments` to
both `UnityBuildPlayerTask` and the `unityBuild` extension. The type of
this property is a simple `Map`. Values provided in the extension will
be configured for all export tasks unless they override this property.

```groovy
unityBuild {
    customArguments = ["value1" : "some value", "aFile": new File("/some/file")]
}
```

```groovy
gradle.taskGraph.whenReady {
  tasks.withType(wooga.gradle.build.unity.tasks.UnityBuildPlayerTask) { t ->
    if(t.buildPlatform == "android") {
      t.customArguments = ['foo':'bar']
    }
  }
}
```

Both key and values of the map will be converted to `String` during
encoding. So custom types with a `toString` implementation are possible.
Be aware that string values with newlines or other special characters
could execute undifined behavior and is also untested at the moment.

Changes
=======

* ![ADD] custom arguments support for `UnityBuildPlayerTask`