Add ExportArchive task
Description
============

This patch adds the second most important xcodebuild functionality, the
ability to create an `ipa` file from an exported `xcarchive`.
The task is based on the new added abstract type
`AbstractXcodeArchiveTask`. This new type shares the majority of
properties with `XcodeArchive` and `ExportArchive`. I also changed the
test setup to reflect this.

Exporting an unsigned ipa is not possible with `xcodebuild` because that
is exactly what the command is for. Create a signed `ipa` package to
upload apple or use to install a local version.

I added tests to verify the creation of an `ipa` file but guarded this
with an `@Requires({ env.TEST_TEAM_ID })` statement.
`TEST_TEAM_ID` should refer to a valid team for which both signing
certificate and provisioning profile for appID
`net.wooga.xcodebuildPluginTest` is awailable on the system that runs
the tests. I used personal signing to test this and the teamID was the
only setting needed.

Changes
=======

* ![ADD] `ExportArchive` task