Add file and pretty console logging (#77)
Description
===========

This patch builds on top of the last addition and prepares a foundation
for future tasks with some common settings for logging and console
printing.

ConsoleSettings
---------------

The console settings config object controls the logoutput of `xcodebuild`
to the standart output. The tool is by nature very verbose and does't allow
any verbosity levels to be set except to quiten it completely. In the apple
development szene is a tool available to help out with that [xcpretty]. It's
main use is to transform the output of `xcodebuild` and format it in a shorter
more human readable format. It also has formatters to reformat unit test output
into different formats. The tool is written in ruby which doesn't help us
much so I ported it over to groovy [xcpretty.groovy]. It is not a 100%
port as code snippet colering is missing and only the `Simple` formatter is
supported. This extra library is added and used when `prettyPrint` is enabled in
the console settings object. The logfile written to disk is not formatted!

The console settings can be set on the plugin level for all `xcodebuild` tasks
or individual for each task.

```groovy
xcodebuild.consoleSettings {
    prettyPrint = true // use xcpretty for console output
    useUnicode = true  // use unicode or ascii output
    colorize = "auto"  // colorize output (always, never, auto)
}
```

The default values are set based on the gradle console settings which can be changed
with the commandline switch `--console=(auto,plain,rich,verbose)` or via gradle property
`-Dorg.gradle.console=(auto,plain,rich,verbose)`

| console setting value | pretty print enabled | unicode enabled    | colorized          |
| --------------------- | -------------------- | ------------------ | ------------------ |
| rich                  | :white_check_mark:   | :white_check_mark: | :white_check_mark: |
| plain                 | :white_check_mark:   |                    |                    |
| verbose               |                      |                    |                    |
| auto                  | :white_check_mark:   | :white_check_mark: | :white_check_mark: |

> because of technical issues with java and gradle it was impossible for me to determine
> a good value if color output should be enabled or not. Normally one check if the desired
> output stream is a tty. This is not possible so I set the `rich` option as the default
> because that is what normal invocation will fallback to.

Extension config settings
-------------------------

I added a few config values to the `XcodeBuildPluginExtion` to set the
default log directory, archive directory and derived data directory.

Each xcodebuild task, at the moment only `XcodeArchive` will bind to
these values by default and append the task name to the log file or as a
sub directory. The common values can be set in the `build.gradle` script
or via environment variable or gradle property.

Changes
=======

* ![ADD] file and pretty console logging

[xcpretty]:        https://github.com/xcpretty/xcpretty
[xcpretty.groovy]: https://github.com/wooga/xcpretty.groovy/