Fix line buffering output for logfile handling (#79)
Description
===========

This is a tempoary fix. The whole outputhandling for xcpretty and
logfile needs to be readressed. I copied the first solution from a
different codebase which used the `LineBufferingOutputStream` from
gradle without need that lines are only written when complete.

Ok the problem at hand that this patch will fix. I need an
`OutputStream` implementation that waits for whole lines to be written
before passing that down to another instance. I used a class from
withing gradle without checking the implementatin details.
The problem with this class is, that it writes out to the connected
handler on the call `flush`. This method can and will be also called
from the outside which can lead to chopped of lines.

I created now my own version of `LineBufferingOutputStream` to solve
this little usability bug. As stated this whole solution is just a
quick fix. I will refactor the whole output handling once more at a
later stage.

Changes
=======

* ![FIX] line buffering output stream for logfile printing