Add ExportArchive task (#78)
* Fix line buffering output for logfile handling

Description
===========

This is a tempoary fix. The whole outputhandling for xcpretty and
logfile needs to be readressed. I copied the first solution from a
different codebase which used the `LineBufferingOutputStream` from
gradle without need that lines are only written when complete.

Ok the problem at hand that this patch will fix. I need an
`OutputStream` implementation that waits for whole lines to be written
before passing that down to another instance. I used a class from
withing gradle without checking the implementatin details.
The problem with this class is, that it writes out to the connected
handler on the call `flush`. This method can and will be also called
from the outside which can lead to chopped of lines.

I created now my own version of `LineBufferingOutputStream` to solve
this little usability bug. As stated this whole solution is just a
quick fix. I will refactor the whole output handling once more at a
later stage.

Changes
=======

* ![FIX] line buffering output stream for logfile printing

* Add ExportArchive task

Description
============

This patch adds the second most important xcodebuild functionality, the
ability to create an `ipa` file from an exported `xcarchive`.
The task is based on the new added abstract type
`AbstractXcodeArchiveTask`. This new type shares the majority of
properties with `XcodeArchive` and `ExportArchive`. I also changed the
test setup to reflect this.

Exporting an unsigned ipa is not possible with `xcodebuild` because that
is exactly what the command is for. Create a signed `ipa` package to
upload apple or use to install a local version.

I added tests to verify the creation of an `ipa` file but guarded this
with an `@Requires({ env.TEST_TEAM_ID })` statement.
`TEST_TEAM_ID` should refer to a valid team for which both signing
certificate and provisioning profile for appID
`net.wooga.xcodebuildPluginTest` is awailable on the system that runs
the tests. I used personal signing to test this and the teamID was the
only setting needed.

Changes
=======

* ![ADD] `ExportArchive` task