PublishArtifact from ExportArchive task
Description
===========

This patch brings a small helper to declare the output of
`ExportArchive` as an gradle publish artifact. The old implementation
`XCodeExportTask` was of type `AbstractArchiveTask` which is by default
an `PublishArtifact`. For the new implementation I did not want to base
of this class because it makes it harder to share common task
properties. Also because the new implementation uses the provider API
which makes it really hard to support both a lazy interface and at the
same time the old static getters. Thats why I opted for a wrapper type.

One can request the `PublishArtifact` from the export task by calling:
`getPublishArtifact()`.

Her is an example usage:

```groovy
configurations {
   archives
}

configurations['default'].extendsFrom(configurations.archives)

project.artifacts {
    archives(${testTaskName}.publishArtifact) {
        it.type = "iOS application archive
    }
}
```

We will need this to be compatible with the current script
implementations.

Changes
=======

* ![ADD] `PublishArtifact` from `ExportArchive` task