Add ArchiveDebugSymbols task type (#81)
Description
===========

The `ArchiveDebugSymbols` task is just a Zip task with no additonal
overloads. I need to have a unique type name for the dynamic
configuration. The task simply zips the `dSYM` directory inside a given
`.xcarchive` as a preparation for deplyment to a symbolication/error
analytics service.

I also add some new configuration that both an `ExportArchive` as well
as an `ArchiveDebugSymbols` task is generated when a task of type
`XcodeArchive` is added.

The generated tasks will take the `XcodeArchive` task name and append
either `ArchiveDSYMS` or `Export`. A later patch will add the
configuration to add these task outputs as project artifacts. So in
theory never should actually call these tasks from the outside.

I adjusted some tests to take the effect of these autogenerated tasks as
a given. I also added some missing tests for earlier patches.

Changes
=======

* ![ADD] `ArchiveDebugSymbols` task type
* ![IMPROVE] autogenerate export and archive debug symbols task