Refactor CreateKeychain task
Description
===========

With the addition of the macos security helper library we
no longer need to write our custom calling implementation in
the gradle plugin. We can leverage the library and simplify
the implementation. I just did that by removing code that
was meanly added as a preperation for other tasks that will follow.
These can now be written in a simple fashion and tested with
the added spock extension.

I also added a missing test which the first impl did not provide.

Changes
=======

* ![IMPROVE] `SecurityCreateKeychain` task implementation