Add SecurityLockKeychain and SecurityUnlockKeychain tasks
Description
===========

This patch adds replacement tasks for LockKeychain task which was able
to lock and unlock. This interface is not very ergonomic as the raw security
commands has a different interface for both actions. The old impl was
oppionated for our internal usecase. The new tasks are bound to the behavior
of the security tool.

Testing these two commands is nearly impossible because the `security` tool
provides no command to test the status of a keychain. The internal macos `security`
contains a function but to write a jni wrapper would be overkill. One could attempt
a general wrapper but the amount of code would not justify it at the moment.

Changes
=======

* ![ADD] `SecurityLockKeychain` task
* ![ADD] `SecurityUnlockKeychain` task