Add SecurityLockKeychain and SecurityUnlockKeychain tasks (#94)
* Add macos keychain spock extension

Description
===========

For better and easier testing this patch adds a special spock
extension library. With this library working with keychains in tests
becomes way easier. Instead of calling create/delete commands manually,
the extension will integrate this automatically.

usage:

```gradle
class ExampleSpec extends Specification {
    @Shared
    @Keychain
    MacOsKeychain testKeychain

    def "create new pull request"() {
        given:
        testKeychain.unlock()

        when:

        then:
    }
}
```

The `@Keychain` extension will return a `MacOsKeychain` object which itself
offers a lot of APIs to manage items in the keychain.

Changes
=======

* ![ADD] macOS keychain spock extension

* Refactor CreateKeychain task

Description
===========

With the addition of the macos security helper library we
no longer need to write our custom calling implementation in
the gradle plugin. We can leverage the library and simplify
the implementation. I just did that by removing code that
was meanly added as a preperation for other tasks that will follow.
These can now be written in a simple fashion and tested with
the added spock extension.

I also added a missing test which the first impl did not provide.

Changes
=======

* ![IMPROVE] `SecurityCreateKeychain` task implementation

* Add SecurityLockKeychain and SecurityUnlockKeychain tasks

Description
===========

This patch adds replacement tasks for LockKeychain task which was able
to lock and unlock. This interface is not very ergonomic as the raw security
commands has a different interface for both actions. The old impl was
oppionated for our internal usecase. The new tasks are bound to the behavior
of the security tool.

Testing these two commands is nearly impossible because the `security` tool
provides no command to test the status of a keychain. The internal macos `security`
contains a function but to write a jni wrapper would be overkill. One could attempt
a general wrapper but the amount of code would not justify it at the moment.

Changes
=======

* ![ADD] `SecurityLockKeychain` task
* ![ADD] `SecurityUnlockKeychain` task