Add reimplementation of KeychainLookupList
Description
===========

This patch brings a reimplementation of the `KeychainLookupList` to
the security library. I replaced all invocations of `SecurityUtil` from
the old class with the new security command versions.

This class will be used to query and alter the keychain lookup list.
The old implementation used to normalize the path passed to the security
cli tool. I reimplemented that also in the new security commands.

Changes
=======

* ![ADD] `KeychainLookupList` class
* ![IMPROVE] normallize keychain path before passing it to `security` cli