Fix generate valid app configs in tests
Description
===========

With a recent change in the plugin `net.wooga.unity`, the plugin
has a stricter parsing defintion for unity asset objects.
This means that we can't simply create a yaml file and let it parse.

The patch adds the normal asset file header for Unity assets so that
the implementation works as expected.

Changes
=======

* ![FIX] generate valid appconfig in tests