Improve keychain cleanup after gradle process kill (#111)
Description
===========

This patch adds a shutdown handler to the `net.wooga.build-io` plugin
to cleanup created build keychains in case of abrupt build process termination.
It uses the `java.lang.Runtime#addShutdownHook` API and registeres a hook
after the `addKeychain` task was executed and removes said hook
when `removeKeychian` was executed. These shutdown hooks run even during normal
termination so it is crucial to only do the work when needed.

I setup a tests that runs a build with a very long task execution in a forked jvm.
The test framework will spawn a new gradle deamon which we will pin to a different
gradle version to find in the process list of macOS. The tests will run with
multiple kill signals to verify the shutdown hook execution.

This whole logic only works for normal kill signals like `TERM` but won't work
for a hard `KILL` signal.

I though about to add more tests to check if the behavior of adding and removing
the shutdown hook works as expected. But in the end it is only important that it
runs correctly during these process shutdown scenarios and that the plugin behaves
normal under normal execution which the other tests already cover.

Changes
=======

* ![IMPROVE] ![MACOS] keychain cleanup after gradle process kill