Improve exported APIs
Description
===========

With the move to gradle 6 we also had to adjust the dependency
components. The `compile` component is no longer available
and one needs to use the `implementation` in most cases.
I moved some dependencies to the `api` component so they can
be directly used by the user of this plugin.

Changes
=======

* ![IMPROVE] exported API