Added new task for new UnityEngine.BuildFromEnvironment entrypoint (#119)
## Description

Created new tasks that communicates with the new UBS 1.2 entrypoint, the `Wooga.UnifiedBuildSystem.Editor.BuildEngine.BuildFromEnvironment` method.  There are two tasks: `UnityBuildEngineTask` and `UnityBuildEnginePlayerTask`. Both are derived from `AbstractUnityBuildEngineTask`, which derives from `atlas-unity` `UnityTask` task. 

`UnityBuildEngineTask` is a basic entrypoint representation. It only needs to know the UBS build that should be triggered and the target version, but also supports a outputPath, a target configuration and secret key/file. The entrypoint method to be used may also be customized. Other arguments to the `-CustomArgs` handler may be passed through the `customArguments` property, which is a list of arguments and support a mix of strings and maps.
 
Basic usage:
```groovy
 task("customExport", type: UnityBuildEngineTask) {
    build = "mandatoryBuildName"
    version = "0.0.1"
 }
```
 
All properties: 
```groovy
 task("customExport", type: UnityBuildEngineTask) {
    build = "mandatoryBuildName"
    version = "0.0.1"
    config = "android" //defaults to empty, which prompts unity to use the UBS default
    exportMethodName = "Unity.Entrypoint.Method" //defaults to Wooga.UnifiedBuildSystem.Editor.BuildEngine.BuildFromEnvironment
    outputPath = "build/output"  //uses project.projectDir as base directory, if path isn't absolute.
    customArguments=["-an-arg", ["-valued-arg":"value"]] //same as ["-an-arg", "-valued-arg=value"]
}
```

`UnityBuildEnginePlayerTask` is targeted at player builds and supports all properties from `UnityBuildEngineTask`, but also supports the `appConfigFile`, `version`,  `versionCode`, `toolsVersion` and `commitHash` properties, with the first two being mandatory, and the others being optional.

Basic usage:
```groovy
 task("customPlayerExport", type: UnityBuildEnginePlayerTask) {
    appConfigFile = file("dir/appConfigFile")
    version = "0.0.1"
}
```


All properties:
```groovy
 task("customPlayerExport", type: UnityBuildEnginePlayerTask) {
    build = "buildName" //defaults to "Player" if not provided
    exportMethodName = "Unity.Entrypoint.Method" //defaults to Wooga.UnifiedBuildSystem.Editor.BuildEngine.BuildFromEnvironment
    outputPath = "build/output"  //uses project.projectDir as base directory, if path isn't absolute.
    customArguments=["-an-arg", ["-valued-arg":"value"]] //same as ["-an-arg", "-valued-arg value"]
    appConfigFile = "dir/appConfigFile" //can be replaced by config
    version = "0.0.1"
    versionCode = "1234" //Optional
    toolsVersion = "0.1.1" //Optional
    commitHash = "a345fc"  //Optional
}
```

The old `UnityBuildPlayerTask` is still present, but will be deprecated in favor of `UnityBuildEnginePlayerTask`.

Finnaly, a disclaimer: I do not have much grasp in the naming part of this job, so feel free to correct or improve names if you think that they aren't good enough :) 

Co-authored-by: Joaquim Neto <joaquim.neto@wooga.net>
Co-authored-by: Manfred Endres <manfred.endres@wooga.net>